# openapi.yaml
openapi: 3.1.0
info:
  title: Model Train Control System API
  description: |
    REST API for managing model trains, edge controllers, and hardware plugins.

    ## Architecture
    - **Central API**: FastAPI backend for train control and configuration
    - **Edge Controllers**: Raspberry Pi devices controlling physical trains
    - **MQTT**: Message broker for real-time command/telemetry

    ## Data Flow
    1. Commands sent via REST API → Published to MQTT
    2. Edge controllers subscribe to MQTT commands
    3. Telemetry from edge controllers → Stored in database
    4. Status retrieved via REST API from database
  version: 0.1.0
  contact:
    name: Model Train Control System
    url: https://github.com/bunchc/model-train-control-system
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: http://central-api:8000
    description: Docker container deployment
paths:
...
  /api/trains:
    get:
      summary: List All Trains
      description: |
        Returns a list of all trains currently configured in the system.
        Note: Currently returns mock data. Will be replaced with database queries.
      operationId: list_trains
      tags:
        - trains
      responses:
        '200':
          description: Successful response with list of trains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Train'
              example:
                - id: "1"
                  name: "Express"
                  description: "High-speed passenger train"
                  model: "Bachmann HO Scale"
                  plugin:
                    name: "motor_controller"
                    config:
                      i2c_address: "0x60"
                - id: "2"
                  name: "Freight"
                  description: "Heavy freight locomotive"
                  model: "Athearn Genesis"
                  plugin:
                    name: "motor_controller"
                    config:
                      i2c_address: "0x61"

    post:
      summary: Register New Train
      description: |
        Register a new train in the system. This endpoint accepts controller_uuid
        in the request body. Edge controllers should prefer using
        POST /api/controllers/{controller_uuid}/trains for clearer ownership.
      operationId: register_train
      tags:
        - trains
      requestBody:
        required: true
        description: Train registration data
        content:
          application/json:
            schema:
              type: object
              required:
                - train_id
                - name
                - controller_uuid
                - plugin
              properties:
                train_id:
                  type: string
                  description: Unique UUID for the train
                  example: "9bf2f703-5ba2-5032-a749-01cce962bcf6"
                name:
                  type: string
                  description: Human-readable train name
                  example: "Express Line Engine"
                controller_uuid:
                  type: string
                  description: UUID of the edge controller managing this train
                  example: "58902c61-2b70-42fe-8542-6b5b404aa998"
                description:
                  type: string
                  description: Optional train description
                  example: "DC Motor on M1 port"
                model:
                  type: string
                  description: Train model name/number
                  example: "DC Motor"
                plugin:
                  type: object
                  required:
                    - name
                  properties:
                    name:
                      type: string
                      example: "dc_motor"
                    config:
                      type: object
                      example:
                        motor_port: 1
      responses:
        '201':
          description: Train registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Train'
        '400':
          description: Controller not found
        '409':
          description: Train ID already exists
  /api/trains/{train_id}/command:
    post:
      summary: Send Command to Train
      description: |
        Publishes a control command to the specified train via MQTT.
        Commands are sent to topic: `trains/{train_id}/commands`

        The edge controller subscribed to this topic will execute the command.
      operationId: send_train_command
      tags:
        - trains
      parameters:
        - name: train_id
          in: path
          required: true
          description: Unique identifier of the train
          schema:
            type: string
          example: "1"
      requestBody:
        required: true
        description: Command payload
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainCommand'
            examples:
              setSpeed:
                summary: Set train speed to 75
                value:
                  action: "setSpeed"
                  speed: 75
              stop:
                summary: Stop the train
                value:
                  action: "stop"
              start:
                summary: Start the train
                value:
                  action: "start"
      responses:
        '200':
          description: Command sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Command sent successfully"
        '400':
          description: Failed to send command
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Failed to send command"
  /api/trains/{train_id}/status:
    get:
      summary: Get Train Status from Database
      description: |
        Retrieves the latest status for a train from the database.
        Status data is periodically updated by edge controllers via MQTT.
      operationId: get_train_status
      tags:
        - trains
      parameters:
        - name: train_id
          in: path
          required: true
          description: Unique identifier of the train
          schema:
            type: string
          example: "1"
      responses:
        '200':
          description: Train status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainStatus'
        '404':
          description: Train status not available
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Train status not available"

  # ============================================================================
  # Configuration Endpoints
  # ============================================================================
...
components:
  schemas:
...
    TrainPlugin:
      type: object
      required:
        - name
      description: Plugin configuration for a specific train
      properties:
        name:
          type: string
          minLength: 1
          description: Name of the plugin to use
          example: "motor_controller"
        config:
          type: object
          additionalProperties: true
          description: Train-specific plugin configuration
          default: {}
          example:
            i2c_address: "0x60"
            default_speed: 50

    # ========================================================================
    # Train Schemas
    # ========================================================================
    Train:
      type: object
      required:
        - id
        - name
        - plugin
      properties:
        id:
          type: string
          minLength: 1
          description: Unique train identifier
          example: "train-001"
        name:
          type: string
          minLength: 1
          description: Human-readable train name
          example: "Express Locomotive"
        description:
          type: string
          nullable: true
          description: Train description
          example: "High-speed passenger train"
        model:
          type: string
          nullable: true
          description: Train model/manufacturer
          example: "Bachmann HO Scale"
        plugin:
          $ref: '#/components/schemas/TrainPlugin'
    TrainStatus:
      type: object
      required:
        - train_id
        - speed
        - voltage
        - current
        - position
      description: Real-time train telemetry and status
      properties:
        train_id:
          type: string
          minLength: 1
          description: Train identifier
          example: "train-001"
        speed:
          type: integer
          minimum: 0
          maximum: 100
          description: Current speed (0-100)
          example: 50
        voltage:
          type: number
          format: float
          minimum: 0
          description: Battery/track voltage in volts
          example: 12.3
        current:
          type: number
          format: float
          minimum: 0
          description: Current draw in amperes
          example: 0.8
        position:
          type: string
          minLength: 1
          description: Current track position/section
          example: "section_A"
    TrainCommand:
      type: object
      description: Command to send to a train
      properties:
        action:
          type: string
          description: Command action type
          enum:
            - setSpeed
            - start
            - stop
            - forward
            - reverse
          example: "setSpeed"
        speed:
          type: integer
          minimum: 0
          maximum: 100
          description: Target speed (required for setSpeed action)
          nullable: true
          example: 75
        direction:
          type: string
          description: Movement direction
          enum:
            - forward
            - reverse
          nullable: true

    # ========================================================================
    # Edge Controller Schemas
    # ========================================================================
...